update on styles.css & index.html
I removed the @import url link for the font family and instead inserted a link on html to enhance performance

I also tried as much as possible to match the h1 and p elements close to the design given in frontend mentor challenge. I know with time I'll get better with this.